cleanup, deleted overlay096 dir and moved the last files into struct_defs
