+++ include/overlay094/gts_application_state.h
@@ -6,13 +6,12 @@
 
 #include "struct_decls/wi_fi_list.h"
 #include "struct_defs/gts_player_data.h"
+#include "struct_defs/world_exchange.h"
 
 #include "applications/pokemon_summary_screen/main.h"
 #include "overlay006/struct_npc_trade_animation_template.h"
 #include "overlay061/struct_ov61_0222C3B0.h"
 #include "overlay094/struct_ov94_02242AAC.h"
-#include "overlay096/struct_world_exchange_trainer.h"
-#include "overlay096/struct_world_exchange_trainer_error.h"
 
 #include "bg_window.h"
 #include "evolution.h"
