--- include/overlay096/struct_world_exchange_trainer_error.h
@@ -1,22 +1,0 @@
-#ifndef POKEPLATINUM_STRUCT_WORLD_EXCHANGE_TRAINER_ERROR_H
-#define POKEPLATINUM_STRUCT_WORLD_EXCHANGE_TRAINER_ERROR_H
-
-enum WorldExchangeValidationError {
-    WE_VALIDATION_ERROR_VALID,
-    WE_VALIDATION_ERROR_FAILED_1,
-    WE_VALIDATION_ERROR_FAILED_2,
-};
-
-enum WorldExchangeSystemError {
-    WE_SYSTEM_ERROR_NONE,
-    WE_SYSTEM_ERROR_DWC_1,
-    WE_SYSTEM_ERROR_DWC_2,
-    WE_SYSTEM_ERROR_DWC_3,
-};
-
-typedef struct WorldExchangeTrainerError {
-    u32 validationError;
-    u32 systemError;
-} WorldExchangeTrainerError;
-
-#endif // POKEPLATINUM_STRUCT_WORLD_EXCHANGE_TRAINER_ERROR_H
