+++ include/wifi_battle_tower/http.h
@@ -2,10 +2,8 @@
 #define POKEPLATINUM_WIFI_BATTLE_TOWER_HTTP_H
 
 #include "struct_defs/wifi_battle_tower_data.h"
-
-#include "overlay096/struct_wifi_player_profile.h"
-#include "overlay096/struct_world_exchange_trainer.h"
-#include "overlay096/struct_world_exchange_trainer_error.h"
+#include "struct_defs/wifi_player_profile.h"
+#include "struct_defs/world_exchange.h"
 
 void BattleTowerHttp_Init(s32 profileId, u64 friendKey);
 void BattleTowerHttp_Update(void);
