+++ include/wifi_battle_tower_save.h
@@ -4,9 +4,9 @@
 #include <nitro/rtc.h>
 
 #include "struct_defs/struct_0202D080.h"
+#include "struct_defs/wifi_player_profile.h"
 #include "struct_defs/wifi_battle_tower_data.h"
 
-#include "overlay096/struct_wifi_player_profile.h"
 #include "overlay104/frontier_data_transfer.h"
 
 #include "easy_chat_sentence.h"
