+++ src/wifi_battle_tower/http.c
@@ -5,11 +5,10 @@
 #include <string.h>
 
 #include "struct_defs/wifi_battle_tower_data.h"
+#include "struct_defs/wifi_player_profile.h"
+#include "struct_defs/world_exchange.h"
 
 #include "http/http.h"
-#include "overlay096/struct_wifi_player_profile.h"
-#include "overlay096/struct_world_exchange_trainer.h"
-#include "overlay096/struct_world_exchange_trainer_error.h"
 #include "wifi_battle_tower/screen.h"
 
 #include "enums.h"
